fix to filter only worker nodes
